requirements.txt
fastapi==0.95.1
uvicorn==0.18.3
tensorflow==2.11.0
numpy==1.23.5
pillow==9.1.0
